practice_1/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.7",
    "greenlet>=3.3.1",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[tool.hatch.build]
    dev-mode-exact = true

[dependency-groups]
    dev = ["editables>=0.5"]
